+++ Source/WebCore/bindings/scripts/IDLAttributes.json
@@ -565,6 +565,9 @@
         "SkipVTableValidation": {
             "contextsAllowed": ["interface"]
         },
+        "StandaloneConstructorAttributes": {
+            "contextsAllowed": ["interface"]
+        },
         "SuppressToJSObject": {
             "contextsAllowed": ["interface"]
         },
